added initial_state to src/index.js
